split FV constants & logging utils
