+++ src/server/verification/api/__tests__/__util__/index.js
@@ -1,7 +1,8 @@
 import { upperFirst, toLower } from 'lodash'
 
+import { enrollmentIdFields } from '../../../utils/logger'
+
 import {
-  enrollmentIdFields,
   failedLivenessMessage,
   failedEnrollmentMessage,
   failedMatchMessage,
@@ -10,7 +11,7 @@ import {
   duplicateFoundMessage,
   alreadyEnrolledMessage,
   successfullyEnrolledMessage
-} from '../../../utils'
+} from '../../../utils/constants'
 
 export default zoomServiceMock => {
   const serviceErrorMessage = 'Request failed with status code 500'
