+++ src/server/verification/processor/provider/ZoomProvider.js
@@ -2,13 +2,15 @@
 import { omit, once, omitBy, bindAll } from 'lodash'
 
 import initZoomAPI from '../../api/ZoomAPI'
+
 import {
-  faceSnapshotFields,
   ZoomAPIError,
   duplicateFoundMessage,
   successfullyEnrolledMessage,
   alreadyEnrolledMessage
-} from '../../utils'
+} from '../../utils/constants'
+
+import { faceSnapshotFields } from '../../utils/logger'
 import logger from '../../../../imports/logger'
 
 import { type IEnrollmentProvider } from '../typings'
